backend/requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework